bug fix
06:46 PM
25-07-2021
the earlier version had a big shortcoming, that it could not identify streaking cells, that is when two adjacent cells are marked, the third cell gets neglected by the complex computer player. it only took into account two non adjacent cells being filled with the middle one empty, like fill-in-the-blank. This version has got that updated and now it works better.